Send to many mails

+++ services/PythonServices/MailSender/helpers/dropbox_files.py
@@ -4,32 +4,42 @@ import tempfile
 import shutil
 from configs.settings import SettingsConfig
 from configs.secrets import SecretsConfig
+from .logger import Logger
 
 
 class DropboxFiles:
     def __init__(self):
         self.__dbx = dropbox.Dropbox(SecretsConfig().dropbox_token)
         self.__temp_dir = os.path.join(tempfile.gettempdir(), "mail_sender")
+        self.__logger = Logger().get_logger()
 
     def download_files(self):
         self.remove_temp_dir()
+        self.__logger.info("Downloading files from " + SettingsConfig().get().dropbox_invoices_dir)
+
         files = self.__get_files(SettingsConfig().get().dropbox_invoices_dir)
+
         for file in files:
             output = os.path.join(self.__temp_dir, file.path_display[1:])
             if not os.path.exists(os.path.dirname(output)):
                 os.makedirs(os.path.dirname(output))
 
+            self.__logger.info("Downloading file " + file.path_display + " to " + output)
             self.__dbx.files_download_to_file(output, file.path_display)
-            print(file.name)
 
     def remove_temp_dir(self):
-        try:
-            shutil.rmtree(self.__temp_dir)
-        except:
-            print('Error while deleting directory')
+        if os.path.isdir(self.__temp_dir):
+            self.__logger.info("Removing temp dir " + self.__temp_dir)
+            try:
+                shutil.rmtree(self.__temp_dir)
+            except:
+                self.__logger.warning("Error while deleting directory")
 
     def __exit__(self, type, value, traceback):
         self.remove_temp_dir()
+
+    def __enter__(self):
+        return self
 
     def __get_files(self, dir):
         files = []
